+++ KnighthoodProject/Assets/Scripts/Weapons/Attack.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Attack", menuName = "ScriptableObjects/Attack")]
 public class Attack : ScriptableObject
 {
-    [field: SerializeField] public byte damage {  get; private set; }
-    [field: SerializeField] public byte armourPen {  get; private set; }
+    [field: SerializeField] public int damage {  get; private set; }
+    [field: SerializeField] public int armourPen {  get; private set; }
     [field: SerializeField] public bool light {  get; private set; }
 }
